requirements.txt
aniso8601==9.0.1
autopep8==1.6.0
camelcase==0.2
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycodestyle==2.8.0
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
Werkzeug==2.1.2
